better diagnostic

+++ Backend/Threads/Thread.h
@@ -20,18 +20,18 @@ class Thread {
    std::thread m_thread;
    std::queue<Data> m_pixels;
 
-   const unsigned int m_id;
+   const size_t m_id;
    const Camera& m_camera;
    const Hittable& m_world;
 
  public:
-   API Thread(const unsigned int id, const Camera& camera, const Hittable& world) noexcept;
+   API Thread(const size_t id, const Camera& camera, const Hittable& world) noexcept;
    API ~Thread();
 
    API void Add(Pixel& pixel, const unsigned int i, const unsigned int j);
    API void StartRender();
 
-   [[nodiscard]] API unsigned int GetId() const noexcept;
+   [[nodiscard]] API size_t GetId() const noexcept;
    [[nodiscard]] API bool RenderDone() const noexcept;
    [[nodiscard]] API size_t ItemsRemaining() const noexcept;
